Extract QuoteCard to a separate module
