+++ app/feature/feature-movingflow/build.gradle.kts
@@ -28,6 +28,7 @@ dependencies {
   implementation(projects.coreUiData)
   implementation(projects.dataContractAndroid)
   implementation(projects.dataProductVariantAndroid)
+  implementation(projects.uiTiersAndAddons)
   implementation(projects.dataProductVariantPublic)
   implementation(projects.designSystemHedvig)
   implementation(projects.moleculeAndroid)
